// web-ui/client/gen/sales/product.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Sales Cookbook API
 * Sales Cookbook API Documentation - A comprehensive sales management system
 * OpenAPI spec version: 1.0.0
 */

export interface Product {
  readonly id: number;
  /** @maxLength 255 */
  name: string;
  description?: string;
  /** @pattern ^-?\d{0,10}(?:\.\d{0,2})?$ */
  price: string;
  /** @maxLength 10 */
  currency?: string;
  is_active?: boolean;
}